CMD script to configure an 100x150 adesive tag.
Here's the situation:

In my job, we needed to configure a printer to print tags so we can paste them into his respectives pallets.

Windows 10 have yet an (hard) access to Internet Explorer, this one that have an module called "Page Setup". This module let us to apply the correct margin and sizes to the tag, so we can print it from the right way.

Althroug that, we have Windows 11...
Microsoft had the geniusly idea to remove the native access to IE in his new OS, and added an "IE Module" at Microsoft Edge. For the surprise of nobody, doesn't work very well...

After days of search and dumb ideas, i think in the basics and applyed some registers at "Regedit", and it works!

Script utility:

- Configure 100x150 adesive tags for printers (I used an Elgin L42 PRO)

Hope that this script helps you.

Good work!